// packages/Common/Types/RollingTime/RollingTime.ts
enum RollingTime {
  Past1Minute = "Past 1 Minute",
  Past5Minutes = "Past 5 Minutes",
  Past10Minutes = "Past 10 Minutes",
  Past15Minutes = "Past 15 Minutes",
  Past30Minutes = "Past 30 Minutes",
  Past1Hour = "Past 1 Hour",
  Past2Hours = "Past 2 Hours",
  Past3Hours = "Past 3 Hours",
  Past6Hours = "Past 6 Hours",
  Past12Hours = "Past 12 Hours",
  Past1Hours = "Past 1 Day",
  Past2Days = "Past 2 Days",
  Past3Days = "Past 3 Days",
  Past7Days = "Past 7 Days",
  Past14Days = "Past 14 Days",
  Past30Days = "Past 30 Days",
  Past60Days = "Past 60 Days",
  Past90Days = "Past 90 Days",
  Past180Days = "Past 180 Days",
  Past365Days = "Past 365 Days",
}

export default RollingTime;